CallTree: show offsets after a symbol name.
tarmac-profile indexes its records by the first address it sees being
executed within a function. So if a trace file starts in mid-function,
and the same function is called again later, there will be one record
for calls to the start of the function, and a separate one for the
case where we started in the middle. But they were both listed in the
output with the same symbol name, confusingly.

Now, when an address is translated into a symbol name, it gets an
offset suffix like '+ 0x1234' if the address turns out not to be the
very start of the function. This is already how symbols are annotated
in the GUI browsers.

This affects both tarmac-calltree and tarmac-profile, because both use
the internal CallTree class.