Support a namespace suffix on Fast Models memory events.
I was just sent a Fast Models Tarmac snippet in which the physical
address field in an MR8 line had an "_NS" suffix. We already supported
this for addresses seen in instruction lines, but this was the first
time I've seen it in a memory read or write. Easily fixed!