Fix accidental ncurses dependence on Unix.
This should allow tarmac-browser to build on a Unix system which has a
curses library that is not ncurses. Previously it would have failed
when browser/curses.cpp tried to `#include <ncurses.h>` and didn't
find it.

The curses-based browser was always intended to work with vanilla
curses, not just ncurses. (Proof: it compiles against pdcurses on
Windows.) So it _should_ work with any curses library that cmake's
FindCurses module can put its hands on. But I hard-coded the directive
`#include <ncurses.h>` in the source file long before I wrote this
cmake setup, and was ignoring the variables retured by FindCurses that
says which curses header file name it found.

Now we use those variables, instead of the OS type, to decide which
header to include. Therefore I also have to manually define one of
them for the pdcurses Windows build, which doesn't run cmake's
FindCurses at all.