Initial impl of detecting CPU exceptions in a trace file.
No functional change yet, except in the diagnostic output of
'parsertest'.

This adds a fifth overload of got_event() in the ParseReceiver base
class. EXC or E lines that describe CPU exceptions are no longer
classed as TextOnlyEvent: they're now ExceptionEvent.

At present the only nontrivial implementation of the new function is
in 'parsertest', which changes its diagnostic output to prove it's
recognised a CPU exception. In the next commit, I'll do something more
interesting with it.

Currently, the only information we parse out of a CPU exception line
is that it _is_ a CPU exception: we don't detect the exception type,
or the PC at which it occurred, or anything else. I don't currently
have a large enough corpus of test input to be sure of the right way
to extract that information, and also, I don't currently have any good
ideas for what the tools could do with the information if it was
available.