Make it easy to find ExceptionEvents in the trace index.
This enables a search tool to find the next or previous CPU exception
from a given line of the trace. However, no tool actually uses it yet.
That's left for the next commit.

Rather than the heavyweight approach of inventing a whole new tree
structure in the index, I've arranged to index exception events in the
existing by-PC tree, by inventing a bogus PC value that means 'CPU
exception'. This value isn't set as curr_pc in any SeqOrderNode; it's
_just_ an extra key in the by-PC tree.

This is an index file format change, so I've bumped the magic number.