Support SVE register updates in FM-style Tarmac.
The test cases for this include some actual Fast Models output, and
also output from another model implementation that very closely mimics
FM's Tarmac style.

In the phony address space we use to represent register contents in
the TTU index file, I've reserved 256 bytes per z-register, since the
largest architectural size of an SVE vector is specified in the ArmARM
to be 2048 bits.

This makes the addressing of d- and s-registers in AArch32 more
complicated. When the q-registers were contiguous in our phony address
space, calculating the location of a d- or s-register was very simple,
because each q-register overlaps the obvious number of d- or s-regs
for its size, so we could treat the d- and s-regs as being contiguous
in the address space too, overlapping the start of the q-space. But
now the q-regs must be spaced apart so that each one overlaps the low
end of an SVE z-register, so we must explicitly divide the d- and
s-register index into a part identifying a q-reg and a part
identifying a location within that q-reg.