Keep track of the largest SVE register update seen.
The whole point of SVE vector registers is that their length can vary
with the implementing CPU, and optionally also with system
configuration (you can deliberately run your SVE program with shorter
vectors than the hardware supports). So if you're reading a trace
generated from an SVE system with relatively short vector registers,
like 256 bits or even the minimum permitted 128, then it would be a
shame if the trace browser insisted on showing 2048 bits for every
vector register, with the vast majority of them in the "don't know"
state.

To avoid this, the indexer tracks the longest SVE register update it's
seen in the whole trace, and writes the result into a header field.