TarmacUtilityBase: factor out get_parse_params().
Similarly to the previous commit: the code which makes a ParseParams
to pass to the Tarmac parser was inside updateIndexIfNeeded(), where
you can only use it if you're also indexing. I'm about to want to
parse without indexing at all, so that code needs to be available from
elsewhere too. Move it into its own method.