New utility 'tarmac-truncate'.
When generating a trace of a program that ends up in an endless loop,
I often pipe the Tarmac output through ad-hoc shell commands to make
it stop at the point where the boring loop begins, so that
tarmac-browser doesn't waste a lot of time and space indexing the part
of the trace nobody cares about, and I can debug more usefully what
happened in the interesting part. This tool supersedes those ad-hoc
commands, so I don't have to make up an appropriate one each time.

Previously I might have used 'sed /xxxxxxxx/q', for example, which
reads input until it sees the pattern xxxxxxxx and then quits, writing
nothing further to the output file and making the Tarmac producer
receive SIGPIPE or EPIPE. The pattern xxxxxxxx might match the
instruction encoding for a 'branch to yourself' tight loop, or a CPU
exception notification, or a vector address, or whatever was
convenient in the particular situation.

The idea is that tarmac-truncate should recognise _all_ the common
patterns in Tarmac that indicate that an unproductive loop is about to
begin. Then I don't have to run the program once without filtering the
trace, see which way it goes wrong, and write a sed command to match
that. I can just pipe through tarmac-truncate in the first place.