CI: add '--output-on-failure' to all the ctest commands.
If a ctest test fails in CI, it's frustrating to have the log of the
workflow run just print the test name with "***Failed", followed by
the pathname of LastTest.log where you would be able to see the
details if only the CI had shown you the contents.

This is particularly annoying if the problem is hard to reproduce,
e.g. intermittent, or only fails on a platform that isn't your native
one.

The simplest answer seems to be to add the '--output-on-failure'
option, so that each failing test will come with details.