Fix uninit variable bug at start of call/return analysis.
The code run for each traced instruction to do call/return analysis
checks the current pc and lr against the member variables
`expected_next_pc` and `expected_next_lr`: the former to detect
transfers of control in the first place, and the latter as one of the
conditions for treating one as a potential call.

At the start of indexing, we don't have any useful value in either of
those variables. `expected_next_pc` field is deliberately initialised
to a known-invalid PC as a sentinel value, and `expected_next_lr` was
accidentally not initialised at all. This caused the indexer to
consider an instruction as a branch when it wasn't (because it checked
against `expected_next_pc` without noticing the value was semantically
None), and then to _intermittently_ consider it as a potential call
depending on what was in the uninitialised `expected_next_lr`.

Now we check both against KNOWN_INVALID_PC before trying to compare
anything else with them.

(Spotted by one of our CI runs – but only the one on macOS Ventura,
where I can't conveniently retest. So the previous commit adding
--output-on-failure allowed me to find out what the problem was.)