GUI browser: substitute branch targets in clipboard data.
The 'substitute branch targets' option causes branch targets like
{pc}+constant to be replaced by symbols from the ELF file in display.
But if you copy-pasted from the trace window into the clipboard, you'd
get the _original_ text from the Tarmac file, which isn't what you
expected. Now you paste the same text you selected.

This is partly just to be more consistent and less confusing, and also
because if you've just found your way to the important part of the
trace, you might specifically _want_ to copy the name of the function
something turns out to be calling, so that you can paste it into a
command that will look it up in the original source code!