GUI browser: fix window resizing.
Previously, each type of TextViewWindow would open at its preferred
size, but (a) wouldn't be resizable _smaller_ than that, and (b) if
resized larger, the contained TextViewCanvas would grow horizontally
but not vertically, in spite of the wxSizerFlags().Expand() in the
code that set up the resizing.

Turns out that simply not trying to use a wxBoxSizer gets most of the
way to this working properly. Calling Fit() on the
TextViewWindow (inherited from wxFrame) means we still calculate the
overall window's size appropriately to the TextViewCanvas's starting
size, but after that, resizing works in all directions - smaller and
bigger, the canvas resizes both horizontally and vertically.

This improved resize handling also affects memory windows: you can now
resize those larger to show more 16-byte rows of memory.

A possible further improvement would be to add some adaptive layout
code, so that register displays could make best use of whatever window
size they're given. For example, the AArch32 core registers currently
display in three columns, no matter how you resize the window; if you
resized it narrower but taller, it would be nice to switch the layout
to two columns or just one, so that everything still fitted. (Indeed,
someone might do that _on purpose_ if they preferred a one-column
layout.)

Similarly, I could imagine it being potentially useful to vary the
number of bytes per line in memory views in response to window
resizing. But that probably has to come with some kind of "common
sense" constraint like keeping it a power of 2. A 13-bytes-per-line
hex dump is helpful in _very_ few circumstances!