tarmac-gui-browser: fix location update when unfolding by mouse.
If the current-position indicator line in the GUI browser is just
after a folded-up function call, and you click the boxed + sign to the
left of the call to unfold it, the function call is unfolded in such a
way as to keep the indicator line after the call, ready to step
forward through the code executed inside the function. But the
'current trace line' indicator in the toolbar was not updated, and
neither were the contents of any linked windows like the registers.

But this all worked fine if you do the same unfolding action using the
right arrow key. So it was just a matter of finding the operation that
happens in the arrow-key case that was missing in the click: a call to
update_location.