Support Fast Models style memory lines with 'xxxx' contents.
These can appear in the output of a Cortex-M3 RTL simulator. At
present I've just made the code handle the case where the _whole_
contents of the memory write is 'xxxxxxxx', which doesn't take much
code. I don't know if it's possible for the trace to show a
_partially_ unspecified memory store (like '0011xx33' or something);
if it is, this commit isn't enough to support that.