[Image] Add support for segments.
The segments, as in the Elf parlance, provide a view of the program's
memory at execution time: where the memory regions are located in memory,
how big they are and their attributes (readable, writable, executable).

This information can be used by when analyzing the memory accesses from
a tarmac trace, without resorting to an external library to parse the
Elf content.

This patch does not add support for grabbing information about the
sections. That would be easy to add, but as there is no immediate use of
it, let's defer this to when it will be needed.