[TarmacUtility] Refactor Image loading and checking.
The Image is now loaded and checked (if an image filename was provided)
before postProcessOptions is invoked. This is a change, but certainly a
desirable one as this allows postProcessOptions from the derived
classes to use the image.

An additional benefit of this refactoring: the endianness consistency
checking is now done for TarmacUtilityMT. It was previously missing from
TarmacUtilityMT::postProcessOptions.

Last, I have also refactored TarmacUtilityBase::load_image() to improve
its readability, and made it a private method as it's not needed outside
of the TarmacUtility base class really.