[tarmac-vcd] Fix how multiple memory accesses are represented in the output VCD file.
In a tarmac trace, instructions comes with several other events:
register updates and memory accesses. Each register is accessed in
parallel so this works fine in the VCD representation that we have (we
have one signal per register). On the other hand, several memory
accesses can happen with the same instruction (e.g. LDM/STM) which are
represented in the tarmac trace as all happening at the same whereas in
the VCD, we have a single MemAddr / MemData busses.

Previously, this was dealt with by adding a 1ns delay in between the
memory accesses, but this had the side effect of making time goes beyond
that of the next instruction when using the timestamps from the tarmac
trace. This was caught by the code and reported with a not so useful
error message.

The solution implemented is to use a much finer timescale (ps instead of
ns) and use those 1000ps to spread the memory accesses (if any). An
assumption here is that there is a time difference of 1 (ns) in between
instructions which does not seem to hold with some tarmac trace
generators (some simulators / vcd writers seems not configured with the
proper timescale). In this case, when using the tarmac trace timestamp,
all bets are off and the outcome will depend on the VCD viewer. GTKWave
will happily overlay the instructions at the same time (which is probably
not was is wanted), while some others may behave differently (only print
the last instruction) or be more creative. This case is reported by
tarmac-vcd and the solution is *not* to use tarmac trace timestamps (the
default).

Update the en_GB@wide and fr translations while being there.